requirements.txt
appdirs>=1.4.4
asgiref>=3.3.1
autopep8>=1.5.6
beautifulsoup4>=4.9.3
bootstrap4>=0.1.0
crontab>=0.22.9
distlib>=0.3.1
dj-database-url>=0.5.0
Django>=3.1.7
django-bootstrap3>=14.2.0
django-bootstrap4>=2.3.1
django-common-helpers>=0.9.2
django-cron>=0.5.1
django-filter>=2.4.0
django-guardian>=2.3.0
django-heroku>=0.3.1
django-jet>=1.0.8
django-suit>=0.2.28
djangorestframework>=3.12.2
filelock>=3.0.12
gunicorn>=20.1.0
Markdown>=3.3.4
pathlib>=1.0.1
pbr>=5.5.1
plotly>=4.14.3
psycopg2>=2.8.6
pycodestyle>=2.7.0
Pygments>=2.8.1
python-decouple>=3.4
pytz>=2021.1
PyYAML>=5.4.1
retrying>=1.3.3
six>=1.15.0
soupsieve>=2.2.1
sqlparse>=0.4.1
stevedore>=3.3.0
toml>=0.10.2
Unipath>=1.1
uritemplate>=3.0.1
virtualenv>=20.4.3
virtualenv-clone>=0.5.4
virtualenvwrapper>=4.8.4
whitenoise>=5.2.0

numpy~=1.20.3
scikit-fuzzy~=0.4.2
matplotlib~=3.4.2
paho-mqtt~=1.5.1
httplib2~=0.19.1